janhaq-backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
pydantic
python-dotenv
PyPDF2
python-docx
python-multipart
